distributed-task-manager/services/api-gateway/requirements.txt
Flask
requests
PyJWT
Pydantic
